did some refactorings to make the code easier to follow.
